requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
pydantic==2.12.5
sqlalchemy>=2.0.0
psycopg2-binary
alembic
python-dotenv